h2 and service test cmmt

+++ src/main/java/com/finartz/restaurantapp/service/ItemService.java
@@ -2,7 +2,6 @@ package com.finartz.restaurantapp.service;
 
 import com.finartz.restaurantapp.model.Item;
 import com.finartz.restaurantapp.repository.ItemRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,8 +9,11 @@ import java.util.List;
 @Service
 public class ItemService {
 
-    @Autowired
-    private ItemRepository itemRepository;
+    private final ItemRepository itemRepository;
+
+    public ItemService(ItemRepository itemRepository) {
+        this.itemRepository = itemRepository;
+    }
 
     public Item create(Item item){
         Item save = itemRepository.save(item);
@@ -37,7 +39,12 @@ public class ItemService {
         return null;
     }
 
-    public void deleteById(Long id){
-        itemRepository.deleteById(id);
+    public Item deleteById(Long id){
+        Item item = itemRepository.getById(id);
+        if (item != null) {
+            itemRepository.deleteById(id);
+            return item;
+        }
+        return item;
     }
 }
